# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FinEdge API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /watchlist:
    get:
      operationId: getWatchlist
      tags: [watchlist]
      summary: Get user watchlist
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Watchlist items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WatchlistItem"
    post:
      operationId: addToWatchlist
      tags: [watchlist]
      summary: Add ticker to watchlist
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistInput"
      responses:
        "201":
          description: Added item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistItem"
  /watchlist/{ticker}:
    delete:
      operationId: removeFromWatchlist
      tags: [watchlist]
      summary: Remove ticker from watchlist
      security:
        - clerkAuth: []
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /stocks/quote/{ticker}:
    get:
      operationId: getStockQuote
      tags: [stocks]
      summary: Get current stock quote
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stock quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockQuote"
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
...
    WatchlistItem:
      type: object
      required: [id, userId, ticker, companyName, addedAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        ticker:
          type: string
        companyName:
          type: string
        addedAt:
          type: string
        quote:
          $ref: "#/components/schemas/StockQuote"
    WatchlistInput:
      type: object
      required: [ticker]
      properties:
        ticker:
          type: string
        companyName:
          type: string
    StockQuote:
      type: object
      required: [ticker, price, change, changePercent, previousClose]
      properties:
        ticker:
          type: string
        companyName:
          type: ["string", "null"]
        price:
          type: number
        change:
          type: number
        changePercent:
          type: number
        previousClose:
          type: number
        previousHigh:
          type: number
        previousLow:
          type: number
        volume:
          type: ["integer", "null"]
        avgVolume:
          type: ["integer", "null"]
        atr:
          type: ["number", "null"]
        isHighVolatility:
          type: boolean
        sparkline:
          type: array
          items:
            type: number
...